requirements.txt
asgiref==3.3.1
Django==3.1.6
django-markdownx==3.0.1
Markdown==3.3.3
Pillow==8.1.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
